Unsafely apply a fast transform

Applies a fast transform created by fort() (x) to the columns of a conformable matrix (y), typically equivalent to the use of the %*% operator, but using an unsafe method.

Usage

x %***% y

Arguments

x

An object of class FastTransform, created using fort().

y

A numeric (real) matrix/vector with an appropriate number of rows/elements.

Value

A numeric (real) matrix with the same number of columns as y.

Details

This operator works in a similar way to %*%, but avoids dispatching and does not perform any type of validation of its inputs, in order to reduce overhead when performing repeated operations inside a function on pre-validated inputs.

It is not recommended that this operator is used interactively and/or on non-validated inputs.

See also

Examples

Z <- fort(4, 1024)
Z %*% matrix(1:2, 4, 3) # output is a 1024 by 3 matrix
#>                  [,1]          [,2]          [,3]
#>    [1,]  1.5613544682  1.5613544682  1.5613544682
#>    [2,] -0.3320104579 -0.3320104579 -0.3320104579
#>    [3,]  0.8668842716  0.8668842716  0.8668842716
#>    [4,] -0.7398596545 -0.7398596545 -0.7398596545
#>    [5,] -0.1376359193 -0.1376359193 -0.1376359193
#>    [6,]  1.7237865698  1.7237865698  1.7237865698
#>    [7,]  1.3897231515  1.3897231515  1.3897231515
#>    [8,] -1.9152729029 -1.9152729029 -1.9152729029
#>    [9,]  0.6415221947  0.6415221947  0.6415221947
#>   [10,] -0.5284129688 -0.5284129688 -0.5284129688
#>   [11,]  2.1679802577  2.1679802577  2.1679802577
#>   [12,]  0.5683097854  0.5683097854  0.5683097854
#>   [13,]  0.9667882925  0.9667882925  0.9667882925
#>   [14,] -0.5683727607 -0.5683727607 -0.5683727607
#>   [15,] -1.0479279654 -1.0479279654 -1.0479279654
#>   [16,]  3.1886983541  3.1886983541  3.1886983541
#>   [17,]  0.4394593067  0.4394593067  0.4394593067
#>   [18,]  0.2044191402  0.2044191402  0.2044191402
#>   [19,] -0.6096750580 -0.6096750580 -0.6096750580
#>   [20,]  2.2264269143  2.2264269143  2.2264269143
#>   [21,]  3.0123957473  3.0123957473  3.0123957473
#>   [22,]  1.4782326758  1.4782326758  1.4782326758
#>   [23,]  0.8118464222  0.8118464222  0.8118464222
#>   [24,] -0.1950520311 -0.1950520311 -0.1950520311
#>   [25,]  0.8490572123  0.8490572123  0.8490572123
#>   [26,] -3.4734440566 -3.4734440566 -3.4734440566
#>   [27,]  0.2810872077  0.2810872077  0.2810872077
#>   [28,] -0.3771254591 -0.3771254591 -0.3771254591
#>   [29,] -1.4815310660 -1.4815310660 -1.4815310660
#>   [30,]  1.3326792632  1.3326792632  1.3326792632
#>   [31,] -1.2742775921 -1.2742775921 -1.2742775921
#>   [32,]  0.1777682140  0.1777682140  0.1777682140
#>   [33,]  0.2185260920  0.2185260920  0.2185260920
#>   [34,] -1.2163687114 -1.2163687114 -1.2163687114
#>   [35,]  0.6228953122  0.6228953122  0.6228953122
#>   [36,]  0.2529646011  0.2529646011  0.2529646011
#>   [37,] -2.1043140245 -2.1043140245 -2.1043140245
#>   [38,]  3.8380474673  3.8380474673  3.8380474673
#>   [39,] -2.4498752154 -2.4498752154 -2.4498752154
#>   [40,] -1.8330793286 -1.8330793286 -1.8330793286
#>   [41,]  0.9068227069  0.9068227069  0.9068227069
#>   [42,] -0.9024840440 -0.9024840440 -0.9024840440
#>   [43,]  0.7275018182  0.7275018182  0.7275018182
#>   [44,] -1.3145477135 -1.3145477135 -1.3145477135
#>   [45,] -2.0395530640 -2.0395530640 -2.0395530640
#>   [46,]  1.6364728773  1.6364728773  1.6364728773
#>   [47,]  0.6330070047  0.6330070047  0.6330070047
#>   [48,] -2.9606394303 -2.9606394303 -2.9606394303
#>   [49,]  0.8083671424  0.8083671424  0.8083671424
#>   [50,]  0.2923357460  0.2923357460  0.2923357460
#>   [51,] -1.1102995626 -1.1102995626 -1.1102995626
#>   [52,] -1.2754148978 -1.2754148978 -1.2754148978
#>   [53,] -0.4535699157 -0.4535699157 -0.4535699157
#>   [54,] -2.0216098431 -2.0216098431 -2.0216098431
#>   [55,] -0.6312400499 -0.6312400499 -0.6312400499
#>   [56,] -0.3780396889 -0.3780396889 -0.3780396889
#>   [57,] -0.3597033226 -0.3597033226 -0.3597033226
#>   [58,] -2.0864963418 -2.0864963418 -2.0864963418
#>   [59,]  0.2755661184  0.2755661184  0.2755661184
#>   [60,] -2.5713056260 -2.5713056260 -2.5713056260
#>   [61,] -2.1428644037 -2.1428644037 -2.1428644037
#>   [62,] -0.1873421487 -0.1873421487 -0.1873421487
#>   [63,]  1.7672687067  1.7672687067  1.7672687067
#>   [64,]  0.0710875262  0.0710875262  0.0710875262
#>   [65,]  1.9140477524  1.9140477524  1.9140477524
#>   [66,]  0.5233215242  0.5233215242  0.5233215242
#>   [67,] -2.3342821702 -2.3342821702 -2.3342821702
#>   [68,]  1.6856750817  1.6856750817  1.6856750817
#>   [69,] -0.5699913000 -0.5699913000 -0.5699913000
#>   [70,]  1.2572466741  1.2572466741  1.2572466741
#>   [71,]  0.5881237729  0.5881237729  0.5881237729
#>   [72,] -2.4256097823 -2.4256097823 -2.4256097823
#>   [73,]  0.9202203436  0.9202203436  0.9202203436
#>   [74,] -1.5387447210 -1.5387447210 -1.5387447210
#>   [75,] -0.0104291836 -0.0104291836 -0.0104291836
#>   [76,]  0.8396740390  0.8396740390  0.8396740390
#>   [77,]  2.2552764552  2.2552764552  2.2552764552
#>   [78,] -0.2884830335 -0.2884830335 -0.2884830335
#>   [79,] -1.6577154347 -1.6577154347 -1.6577154347
#>   [80,] -0.6717310682 -0.6717310682 -0.6717310682
#>   [81,]  1.9882907244  1.9882907244  1.9882907244
#>   [82,]  1.0015226282  1.0015226282  1.0015226282
#>   [83,]  0.0725748399  0.0725748399  0.0725748399
#>   [84,] -1.2166789870 -1.2166789870 -1.2166789870
#>   [85,]  1.3986873146  1.3986873146  1.3986873146
#>   [86,]  1.8524652725  1.8524652725  1.8524652725
#>   [87,] -1.0337976814 -1.0337976814 -1.0337976814
#>   [88,]  0.1656457054  0.1656457054  0.1656457054
#>   [89,]  0.6382953580  0.6382953580  0.6382953580
#>   [90,] -3.0884964319 -3.0884964319 -3.0884964319
#>   [91,] -1.4259958269 -1.4259958269 -1.4259958269
#>   [92,] -2.8196294519 -2.8196294519 -2.8196294519
#>   [93,]  1.5391727682  1.5391727682  1.5391727682
#>   [94,]  1.4417293899  1.4417293899  1.4417293899
#>   [95,]  1.5480195197  1.5480195197  1.5480195197
#>   [96,] -2.0876635433 -2.0876635433 -2.0876635433
#>   [97,]  1.5763568064  1.5763568064  1.5763568064
#>   [98,]  0.4113960130  0.4113960130  0.4113960130
#>   [99,] -0.4746246570 -0.4746246570 -0.4746246570
#>  [100,] -1.0675753250 -1.0675753250 -1.0675753250
#>  [101,]  0.2337661640  0.2337661640  0.2337661640
#>  [102,] -2.1976169093 -2.1976169093 -2.1976169093
#>  [103,] -0.2742653864 -0.2742653864 -0.2742653864
#>  [104,]  1.7367123273  1.7367123273  1.7367123273
#>  [105,]  1.0300807222  1.0300807222  1.0300807222
#>  [106,]  0.3614069713  0.3614069713  0.3614069713
#>  [107,]  0.5586422250  0.5586422250  0.5586422250
#>  [108,] -0.0045459702 -0.0045459702 -0.0045459702
#>  [109,] -0.4563865191 -0.4563865191 -0.4563865191
#>  [110,] -0.9174921505 -0.9174921505 -0.9174921505
#>  [111,] -0.0664723090 -0.0664723090 -0.0664723090
#>  [112,] -1.6897189380 -1.6897189380 -1.6897189380
#>  [113,]  1.1241750655  1.1241750655  1.1241750655
#>  [114,] -0.4774490437 -0.4774490437 -0.4774490437
#>  [115,] -1.8865987134 -1.8865987134 -1.8865987134
#>  [116,]  2.7774988520  2.7774988520  2.7774988520
#>  [117,]  1.8179062591  1.8179062591  1.8179062591
#>  [118,]  1.3220224143  1.3220224143  1.3220224143
#>  [119,] -1.5345998739 -1.5345998739 -1.5345998739
#>  [120,] -2.1030821378 -2.1030821378 -2.1030821378
#>  [121,]  2.5946400267  2.5946400267  2.5946400267
#>  [122,] -4.7753677692 -4.7753677692 -4.7753677692
#>  [123,] -0.3836723087 -0.3836723087 -0.3836723087
#>  [124,] -2.2632443070 -2.2632443070 -2.2632443070
#>  [125,] -0.4663961904 -0.4663961904 -0.4663961904
#>  [126,] -2.1355112622 -2.1355112622 -2.1355112622
#>  [127,] -0.9281483502 -0.9281483502 -0.9281483502
#>  [128,]  0.9167997606  0.9167997606  0.9167997606
#>  [129,]  4.6041839635  4.6041839635  4.6041839635
#>  [130,]  2.4280816863  2.4280816863  2.4280816863
#>  [131,]  0.4361540840  0.4361540840  0.4361540840
#>  [132,]  2.1458239306  2.1458239306  2.1458239306
#>  [133,]  2.9002416661  2.9002416661  2.9002416661
#>  [134,] -1.0878517361 -1.0878517361 -1.0878517361
#>  [135,] -2.3466815636 -2.3466815636 -2.3466815636
#>  [136,]  0.0765932579  0.0765932579  0.0765932579
#>  [137,] -0.7458960933 -0.7458960933 -0.7458960933
#>  [138,]  1.0229575382  1.0229575382  1.0229575382
#>  [139,] -1.4404611367 -1.4404611367 -1.4404611367
#>  [140,]  1.4656138020  1.4656138020  1.4656138020
#>  [141,]  2.2989469212  2.2989469212  2.2989469212
#>  [142,]  1.5973241005  1.5973241005  1.5973241005
#>  [143,]  1.9641286704  1.9641286704  1.9641286704
#>  [144,] -0.3357856563 -0.3357856563 -0.3357856563
#>  [145,]  1.9773566057  1.9773566057  1.9773566057
#>  [146,]  0.1287231450  0.1287231450  0.1287231450
#>  [147,] -0.7109526039 -0.7109526039 -0.7109526039
#>  [148,]  1.9910088607  1.9910088607  1.9910088607
#>  [149,]  2.5609953398  2.5609953398  2.5609953398
#>  [150,]  0.1846389978  0.1846389978  0.1846389978
#>  [151,]  2.4080010470  2.4080010470  2.4080010470
#>  [152,] -1.6844947416 -1.6844947416 -1.6844947416
#>  [153,]  0.9448002129  0.9448002129  0.9448002129
#>  [154,] -1.2660624928 -1.2660624928 -1.2660624928
#>  [155,]  0.7908321404  0.7908321404  0.7908321404
#>  [156,]  0.1122207271  0.1122207271  0.1122207271
#>  [157,] -1.8587628317 -1.8587628317 -1.8587628317
#>  [158,] -1.5071530601 -1.5071530601 -1.5071530601
#>  [159,]  1.8739990455  1.8739990455  1.8739990455
#>  [160,]  3.4279757551  3.4279757551  3.4279757551
#>  [161,]  0.6638520569  0.6638520569  0.6638520569
#>  [162,]  0.0482050973  0.0482050973  0.0482050973
#>  [163,] -1.8681684237 -1.8681684237 -1.8681684237
#>  [164,]  0.7814157375  0.7814157375  0.7814157375
#>  [165,]  0.5871628678  0.5871628678  0.5871628678
#>  [166,]  1.6588656773  1.6588656773  1.6588656773
#>  [167,]  0.6844394330  0.6844394330  0.6844394330
#>  [168,] -0.1454636125 -0.1454636125 -0.1454636125
#>  [169,]  1.3276212229  1.3276212229  1.3276212229
#>  [170,]  0.0619293020  0.0619293020  0.0619293020
#>  [171,]  1.0790231890  1.0790231890  1.0790231890
#>  [172,] -1.8032760330 -1.8032760330 -1.8032760330
#>  [173,]  1.8690920410  1.8690920410  1.8690920410
#>  [174,]  0.8774310172  0.8774310172  0.8774310172
#>  [175,] -0.9302953803 -0.9302953803 -0.9302953803
#>  [176,] -1.0516304506 -1.0516304506 -1.0516304506
#>  [177,]  0.2979350013  0.2979350013  0.2979350013
#>  [178,]  0.0774877238  0.0774877238  0.0774877238
#>  [179,]  1.3667734939  1.3667734939  1.3667734939
#>  [180,] -2.2378580767 -2.2378580767 -2.2378580767
#>  [181,]  1.1909175856  1.1909175856  1.1909175856
#>  [182,]  0.0974681958  0.0974681958  0.0974681958
#>  [183,]  0.0210380793  0.0210380793  0.0210380793
#>  [184,]  1.2797856428  1.2797856428  1.2797856428
#>  [185,] -0.9273881962 -0.9273881962 -0.9273881962
#>  [186,]  0.1859988516  0.1859988516  0.1859988516
#>  [187,]  2.0667993880  2.0667993880  2.0667993880
#>  [188,]  2.1246689578  2.1246689578  2.1246689578
#>  [189,] -1.2790586279 -1.2790586279 -1.2790586279
#>  [190,] -0.2392797032 -0.2392797032 -0.2392797032
#>  [191,]  1.8106317160  1.8106317160  1.8106317160
#>  [192,] -2.4902887890 -2.4902887890 -2.4902887890
#>  [193,] -2.0283650108 -2.0283650108 -2.0283650108
#>  [194,]  1.2366519827  1.2366519827  1.2366519827
#>  [195,] -0.4649959637 -0.4649959637 -0.4649959637
#>  [196,] -0.8386175992 -0.8386175992 -0.8386175992
#>  [197,]  0.7275400452  0.7275400452  0.7275400452
#>  [198,]  0.5606813444  0.5606813444  0.5606813444
#>  [199,]  1.5484212996  1.5484212996  1.5484212996
#>  [200,]  4.1495994996  4.1495994996  4.1495994996
#>  [201,]  0.3538982774  0.3538982774  0.3538982774
#>  [202,] -3.9271254191 -3.9271254191 -3.9271254191
#>  [203,] -0.8052943021 -0.8052943021 -0.8052943021
#>  [204,]  3.0247217767  3.0247217767  3.0247217767
#>  [205,]  1.3831006767  1.3831006767  1.3831006767
#>  [206,] -2.9330780345 -2.9330780345 -2.9330780345
#>  [207,] -1.1549787549 -1.1549787549 -1.1549787549
#>  [208,] -1.2516783106 -1.2516783106 -1.2516783106
#>  [209,]  1.9514733444  1.9514733444  1.9514733444
#>  [210,]  0.2379130054  0.2379130054  0.2379130054
#>  [211,] -1.1982540108 -1.1982540108 -1.1982540108
#>  [212,]  1.7346093749  1.7346093749  1.7346093749
#>  [213,]  0.6102780971  0.6102780971  0.6102780971
#>  [214,] -0.0169038270 -0.0169038270 -0.0169038270
#>  [215,]  2.6425404146  2.6425404146  2.6425404146
#>  [216,]  0.2319015951  0.2319015951  0.2319015951
#>  [217,]  0.1306587529  0.1306587529  0.1306587529
#>  [218,]  0.2159798566  0.2159798566  0.2159798566
#>  [219,]  1.8741904597  1.8741904597  1.8741904597
#>  [220,] -1.1373311223 -1.1373311223 -1.1373311223
#>  [221,]  0.2867476303  0.2867476303  0.2867476303
#>  [222,] -3.3167651869 -3.3167651869 -3.3167651869
#>  [223,] -0.6472349037 -0.6472349037 -0.6472349037
#>  [224,]  0.8703245665  0.8703245665  0.8703245665
#>  [225,] -0.6981325003 -0.6981325003 -0.6981325003
#>  [226,]  0.5889908768  0.5889908768  0.5889908768
#>  [227,] -2.3865063159 -2.3865063159 -2.3865063159
#>  [228,] -0.2491391371 -0.2491391371 -0.2491391371
#>  [229,]  0.1036564266  0.1036564266  0.1036564266
#>  [230,]  2.0300800181  2.0300800181  2.0300800181
#>  [231,]  2.7195292403  2.7195292403  2.7195292403
#>  [232,] -0.7839505791 -0.7839505791 -0.7839505791
#>  [233,]  1.0316857345  1.0316857345  1.0316857345
#>  [234,]  0.8326247935  0.8326247935  0.8326247935
#>  [235,] -2.7544539369 -2.7544539369 -2.7544539369
#>  [236,] -0.7711489501 -0.7711489501 -0.7711489501
#>  [237,]  2.1222612217  2.1222612217  2.1222612217
#>  [238,]  0.1947234221  0.1947234221  0.1947234221
#>  [239,] -1.1144195291 -1.1144195291 -1.1144195291
#>  [240,]  0.3352564708  0.3352564708  0.3352564708
#>  [241,] -0.6702600546 -0.6702600546 -0.6702600546
#>  [242,]  1.3753831754  1.3753831754  1.3753831754
#>  [243,]  2.2730284290  2.2730284290  2.2730284290
#>  [244,]  2.6645635789  2.6645635789  2.6645635789
#>  [245,] -2.2237440716 -2.2237440716 -2.2237440716
#>  [246,]  0.6643715724  0.6643715724  0.6643715724
#>  [247,]  0.2111992773  0.2111992773  0.2111992773
#>  [248,] -1.2128635438 -1.2128635438 -1.2128635438
#>  [249,] -0.0903205953 -0.0903205953 -0.0903205953
#>  [250,] -1.7705689144 -1.7705689144 -1.7705689144
#>  [251,] -2.7541505384 -2.7541505384 -2.7541505384
#>  [252,] -0.9075087864 -0.9075087864 -0.9075087864
#>  [253,]  0.1386878779  0.1386878779  0.1386878779
#>  [254,]  0.7363067591  0.7363067591  0.7363067591
#>  [255,] -0.4427593319 -0.4427593319 -0.4427593319
#>  [256,] -1.1156702288 -1.1156702288 -1.1156702288
#>  [257,] -0.0153272238 -0.0153272238 -0.0153272238
#>  [258,] -2.3464453668 -2.3464453668 -2.3464453668
#>  [259,]  0.5681734253  0.5681734253  0.5681734253
#>  [260,] -2.4158498868 -2.4158498868 -2.4158498868
#>  [261,]  0.2753182744  0.2753182744  0.2753182744
#>  [262,]  2.4742493806  2.4742493806  2.4742493806
#>  [263,]  0.7486599784  0.7486599784  0.7486599784
#>  [264,]  0.8948575151  0.8948575151  0.8948575151
#>  [265,] -3.5233052578 -3.5233052578 -3.5233052578
#>  [266,] -1.6848420547 -1.6848420547 -1.6848420547
#>  [267,] -0.9318068783 -0.9318068783 -0.9318068783
#>  [268,] -2.8192831706 -2.8192831706 -2.8192831706
#>  [269,]  0.1098819007  0.1098819007  0.1098819007
#>  [270,]  0.6332697203  0.6332697203  0.6332697203
#>  [271,]  0.5732487283  0.5732487283  0.5732487283
#>  [272,]  1.0387559896  1.0387559896  1.0387559896
#>  [273,]  1.4839524940  1.4839524940  1.4839524940
#>  [274,]  0.7543696625  0.7543696625  0.7543696625
#>  [275,]  0.5061546843  0.5061546843  0.5061546843
#>  [276,] -0.9039528692 -0.9039528692 -0.9039528692
#>  [277,] -0.5478363684 -0.5478363684 -0.5478363684
#>  [278,]  0.1166785754  0.1166785754  0.1166785754
#>  [279,] -0.1267028101 -0.1267028101 -0.1267028101
#>  [280,] -1.1185029037 -1.1185029037 -1.1185029037
#>  [281,] -1.0276025089 -1.0276025089 -1.0276025089
#>  [282,] -1.8025445834 -1.8025445834 -1.8025445834
#>  [283,] -3.0241276767 -3.0241276767 -3.0241276767
#>  [284,]  0.8175535069  0.8175535069  0.8175535069
#>  [285,]  0.7345423752  0.7345423752  0.7345423752
#>  [286,]  2.9807136440  2.9807136440  2.9807136440
#>  [287,] -1.7633297314 -1.7633297314 -1.7633297314
#>  [288,]  1.7820967090  1.7820967090  1.7820967090
#>  [289,] -2.6897532533 -2.6897532533 -2.6897532533
#>  [290,] -1.9267193322 -1.9267193322 -1.9267193322
#>  [291,]  1.3765274312  1.3765274312  1.3765274312
#>  [292,] -2.7473072389 -2.7473072389 -2.7473072389
#>  [293,] -2.0981871884 -2.0981871884 -2.0981871884
#>  [294,] -1.3386250578 -1.3386250578 -1.3386250578
#>  [295,] -1.6495026546 -1.6495026546 -1.6495026546
#>  [296,]  0.7271247626  0.7271247626  0.7271247626
#>  [297,] -0.3760531432 -0.3760531432 -0.3760531432
#>  [298,]  3.0079019940  3.0079019940  3.0079019940
#>  [299,]  0.6339680437  0.6339680437  0.6339680437
#>  [300,] -0.1681017287 -0.1681017287 -0.1681017287
#>  [301,]  0.8392188259  0.8392188259  0.8392188259
#>  [302,]  0.0647900542  0.0647900542  0.0647900542
#>  [303,] -0.2865079113 -0.2865079113 -0.2865079113
#>  [304,]  0.3198018829  0.3198018829  0.3198018829
#>  [305,] -1.4872184252 -1.4872184252 -1.4872184252
#>  [306,] -2.8325578658 -2.8325578658 -2.8325578658
#>  [307,]  1.8658286974  1.8658286974  1.8658286974
#>  [308,]  0.2815269294  0.2815269294  0.2815269294
#>  [309,]  2.0614379277  2.0614379277  2.0614379277
#>  [310,] -1.9565021555 -1.9565021555 -1.9565021555
#>  [311,] -1.2696405655 -1.2696405655 -1.2696405655
#>  [312,] -0.0884776849 -0.0884776849 -0.0884776849
#>  [313,] -0.2629123442 -0.2629123442 -0.2629123442
#>  [314,] -0.9475134768 -0.9475134768 -0.9475134768
#>  [315,]  3.5586860512  3.5586860512  3.5586860512
#>  [316,] -0.7679478028 -0.7679478028 -0.7679478028
#>  [317,] -0.7835007519 -0.7835007519 -0.7835007519
#>  [318,] -3.4300716311 -3.4300716311 -3.4300716311
#>  [319,]  0.3701204013  0.3701204013  0.3701204013
#>  [320,]  0.2574144901  0.2574144901  0.2574144901
#>  [321,]  1.9775510498  1.9775510498  1.9775510498
#>  [322,]  1.4516772647  1.4516772647  1.4516772647
#>  [323,]  0.7143802477  0.7143802477  0.7143802477
#>  [324,] -2.5333232208 -2.5333232208 -2.5333232208
#>  [325,]  1.1493114757  1.1493114757  1.1493114757
#>  [326,]  1.1495537560  1.1495537560  1.1495537560
#>  [327,] -0.6777746473 -0.6777746473 -0.6777746473
#>  [328,]  2.0269313180  2.0269313180  2.0269313180
#>  [329,]  2.5526643573  2.5526643573  2.5526643573
#>  [330,]  0.2309324951  0.2309324951  0.2309324951
#>  [331,]  2.5695464927  2.5695464927  2.5695464927
#>  [332,]  1.5448881424  1.5448881424  1.5448881424
#>  [333,] -0.7863377242 -0.7863377242 -0.7863377242
#>  [ reached getOption("max.print") -- omitted 691 rows ]
Z %***% matrix(1:2, 4, 3) # output is also a 1024 by 3 matrix
#>                  [,1]          [,2]          [,3]
#>    [1,]  1.5613544682  1.5613544682  1.5613544682
#>    [2,] -0.3320104579 -0.3320104579 -0.3320104579
#>    [3,]  0.8668842716  0.8668842716  0.8668842716
#>    [4,] -0.7398596545 -0.7398596545 -0.7398596545
#>    [5,] -0.1376359193 -0.1376359193 -0.1376359193
#>    [6,]  1.7237865698  1.7237865698  1.7237865698
#>    [7,]  1.3897231515  1.3897231515  1.3897231515
#>    [8,] -1.9152729029 -1.9152729029 -1.9152729029
#>    [9,]  0.6415221947  0.6415221947  0.6415221947
#>   [10,] -0.5284129688 -0.5284129688 -0.5284129688
#>   [11,]  2.1679802577  2.1679802577  2.1679802577
#>   [12,]  0.5683097854  0.5683097854  0.5683097854
#>   [13,]  0.9667882925  0.9667882925  0.9667882925
#>   [14,] -0.5683727607 -0.5683727607 -0.5683727607
#>   [15,] -1.0479279654 -1.0479279654 -1.0479279654
#>   [16,]  3.1886983541  3.1886983541  3.1886983541
#>   [17,]  0.4394593067  0.4394593067  0.4394593067
#>   [18,]  0.2044191402  0.2044191402  0.2044191402
#>   [19,] -0.6096750580 -0.6096750580 -0.6096750580
#>   [20,]  2.2264269143  2.2264269143  2.2264269143
#>   [21,]  3.0123957473  3.0123957473  3.0123957473
#>   [22,]  1.4782326758  1.4782326758  1.4782326758
#>   [23,]  0.8118464222  0.8118464222  0.8118464222
#>   [24,] -0.1950520311 -0.1950520311 -0.1950520311
#>   [25,]  0.8490572123  0.8490572123  0.8490572123
#>   [26,] -3.4734440566 -3.4734440566 -3.4734440566
#>   [27,]  0.2810872077  0.2810872077  0.2810872077
#>   [28,] -0.3771254591 -0.3771254591 -0.3771254591
#>   [29,] -1.4815310660 -1.4815310660 -1.4815310660
#>   [30,]  1.3326792632  1.3326792632  1.3326792632
#>   [31,] -1.2742775921 -1.2742775921 -1.2742775921
#>   [32,]  0.1777682140  0.1777682140  0.1777682140
#>   [33,]  0.2185260920  0.2185260920  0.2185260920
#>   [34,] -1.2163687114 -1.2163687114 -1.2163687114
#>   [35,]  0.6228953122  0.6228953122  0.6228953122
#>   [36,]  0.2529646011  0.2529646011  0.2529646011
#>   [37,] -2.1043140245 -2.1043140245 -2.1043140245
#>   [38,]  3.8380474673  3.8380474673  3.8380474673
#>   [39,] -2.4498752154 -2.4498752154 -2.4498752154
#>   [40,] -1.8330793286 -1.8330793286 -1.8330793286
#>   [41,]  0.9068227069  0.9068227069  0.9068227069
#>   [42,] -0.9024840440 -0.9024840440 -0.9024840440
#>   [43,]  0.7275018182  0.7275018182  0.7275018182
#>   [44,] -1.3145477135 -1.3145477135 -1.3145477135
#>   [45,] -2.0395530640 -2.0395530640 -2.0395530640
#>   [46,]  1.6364728773  1.6364728773  1.6364728773
#>   [47,]  0.6330070047  0.6330070047  0.6330070047
#>   [48,] -2.9606394303 -2.9606394303 -2.9606394303
#>   [49,]  0.8083671424  0.8083671424  0.8083671424
#>   [50,]  0.2923357460  0.2923357460  0.2923357460
#>   [51,] -1.1102995626 -1.1102995626 -1.1102995626
#>   [52,] -1.2754148978 -1.2754148978 -1.2754148978
#>   [53,] -0.4535699157 -0.4535699157 -0.4535699157
#>   [54,] -2.0216098431 -2.0216098431 -2.0216098431
#>   [55,] -0.6312400499 -0.6312400499 -0.6312400499
#>   [56,] -0.3780396889 -0.3780396889 -0.3780396889
#>   [57,] -0.3597033226 -0.3597033226 -0.3597033226
#>   [58,] -2.0864963418 -2.0864963418 -2.0864963418
#>   [59,]  0.2755661184  0.2755661184  0.2755661184
#>   [60,] -2.5713056260 -2.5713056260 -2.5713056260
#>   [61,] -2.1428644037 -2.1428644037 -2.1428644037
#>   [62,] -0.1873421487 -0.1873421487 -0.1873421487
#>   [63,]  1.7672687067  1.7672687067  1.7672687067
#>   [64,]  0.0710875262  0.0710875262  0.0710875262
#>   [65,]  1.9140477524  1.9140477524  1.9140477524
#>   [66,]  0.5233215242  0.5233215242  0.5233215242
#>   [67,] -2.3342821702 -2.3342821702 -2.3342821702
#>   [68,]  1.6856750817  1.6856750817  1.6856750817
#>   [69,] -0.5699913000 -0.5699913000 -0.5699913000
#>   [70,]  1.2572466741  1.2572466741  1.2572466741
#>   [71,]  0.5881237729  0.5881237729  0.5881237729
#>   [72,] -2.4256097823 -2.4256097823 -2.4256097823
#>   [73,]  0.9202203436  0.9202203436  0.9202203436
#>   [74,] -1.5387447210 -1.5387447210 -1.5387447210
#>   [75,] -0.0104291836 -0.0104291836 -0.0104291836
#>   [76,]  0.8396740390  0.8396740390  0.8396740390
#>   [77,]  2.2552764552  2.2552764552  2.2552764552
#>   [78,] -0.2884830335 -0.2884830335 -0.2884830335
#>   [79,] -1.6577154347 -1.6577154347 -1.6577154347
#>   [80,] -0.6717310682 -0.6717310682 -0.6717310682
#>   [81,]  1.9882907244  1.9882907244  1.9882907244
#>   [82,]  1.0015226282  1.0015226282  1.0015226282
#>   [83,]  0.0725748399  0.0725748399  0.0725748399
#>   [84,] -1.2166789870 -1.2166789870 -1.2166789870
#>   [85,]  1.3986873146  1.3986873146  1.3986873146
#>   [86,]  1.8524652725  1.8524652725  1.8524652725
#>   [87,] -1.0337976814 -1.0337976814 -1.0337976814
#>   [88,]  0.1656457054  0.1656457054  0.1656457054
#>   [89,]  0.6382953580  0.6382953580  0.6382953580
#>   [90,] -3.0884964319 -3.0884964319 -3.0884964319
#>   [91,] -1.4259958269 -1.4259958269 -1.4259958269
#>   [92,] -2.8196294519 -2.8196294519 -2.8196294519
#>   [93,]  1.5391727682  1.5391727682  1.5391727682
#>   [94,]  1.4417293899  1.4417293899  1.4417293899
#>   [95,]  1.5480195197  1.5480195197  1.5480195197
#>   [96,] -2.0876635433 -2.0876635433 -2.0876635433
#>   [97,]  1.5763568064  1.5763568064  1.5763568064
#>   [98,]  0.4113960130  0.4113960130  0.4113960130
#>   [99,] -0.4746246570 -0.4746246570 -0.4746246570
#>  [100,] -1.0675753250 -1.0675753250 -1.0675753250
#>  [101,]  0.2337661640  0.2337661640  0.2337661640
#>  [102,] -2.1976169093 -2.1976169093 -2.1976169093
#>  [103,] -0.2742653864 -0.2742653864 -0.2742653864
#>  [104,]  1.7367123273  1.7367123273  1.7367123273
#>  [105,]  1.0300807222  1.0300807222  1.0300807222
#>  [106,]  0.3614069713  0.3614069713  0.3614069713
#>  [107,]  0.5586422250  0.5586422250  0.5586422250
#>  [108,] -0.0045459702 -0.0045459702 -0.0045459702
#>  [109,] -0.4563865191 -0.4563865191 -0.4563865191
#>  [110,] -0.9174921505 -0.9174921505 -0.9174921505
#>  [111,] -0.0664723090 -0.0664723090 -0.0664723090
#>  [112,] -1.6897189380 -1.6897189380 -1.6897189380
#>  [113,]  1.1241750655  1.1241750655  1.1241750655
#>  [114,] -0.4774490437 -0.4774490437 -0.4774490437
#>  [115,] -1.8865987134 -1.8865987134 -1.8865987134
#>  [116,]  2.7774988520  2.7774988520  2.7774988520
#>  [117,]  1.8179062591  1.8179062591  1.8179062591
#>  [118,]  1.3220224143  1.3220224143  1.3220224143
#>  [119,] -1.5345998739 -1.5345998739 -1.5345998739
#>  [120,] -2.1030821378 -2.1030821378 -2.1030821378
#>  [121,]  2.5946400267  2.5946400267  2.5946400267
#>  [122,] -4.7753677692 -4.7753677692 -4.7753677692
#>  [123,] -0.3836723087 -0.3836723087 -0.3836723087
#>  [124,] -2.2632443070 -2.2632443070 -2.2632443070
#>  [125,] -0.4663961904 -0.4663961904 -0.4663961904
#>  [126,] -2.1355112622 -2.1355112622 -2.1355112622
#>  [127,] -0.9281483502 -0.9281483502 -0.9281483502
#>  [128,]  0.9167997606  0.9167997606  0.9167997606
#>  [129,]  4.6041839635  4.6041839635  4.6041839635
#>  [130,]  2.4280816863  2.4280816863  2.4280816863
#>  [131,]  0.4361540840  0.4361540840  0.4361540840
#>  [132,]  2.1458239306  2.1458239306  2.1458239306
#>  [133,]  2.9002416661  2.9002416661  2.9002416661
#>  [134,] -1.0878517361 -1.0878517361 -1.0878517361
#>  [135,] -2.3466815636 -2.3466815636 -2.3466815636
#>  [136,]  0.0765932579  0.0765932579  0.0765932579
#>  [137,] -0.7458960933 -0.7458960933 -0.7458960933
#>  [138,]  1.0229575382  1.0229575382  1.0229575382
#>  [139,] -1.4404611367 -1.4404611367 -1.4404611367
#>  [140,]  1.4656138020  1.4656138020  1.4656138020
#>  [141,]  2.2989469212  2.2989469212  2.2989469212
#>  [142,]  1.5973241005  1.5973241005  1.5973241005
#>  [143,]  1.9641286704  1.9641286704  1.9641286704
#>  [144,] -0.3357856563 -0.3357856563 -0.3357856563
#>  [145,]  1.9773566057  1.9773566057  1.9773566057
#>  [146,]  0.1287231450  0.1287231450  0.1287231450
#>  [147,] -0.7109526039 -0.7109526039 -0.7109526039
#>  [148,]  1.9910088607  1.9910088607  1.9910088607
#>  [149,]  2.5609953398  2.5609953398  2.5609953398
#>  [150,]  0.1846389978  0.1846389978  0.1846389978
#>  [151,]  2.4080010470  2.4080010470  2.4080010470
#>  [152,] -1.6844947416 -1.6844947416 -1.6844947416
#>  [153,]  0.9448002129  0.9448002129  0.9448002129
#>  [154,] -1.2660624928 -1.2660624928 -1.2660624928
#>  [155,]  0.7908321404  0.7908321404  0.7908321404
#>  [156,]  0.1122207271  0.1122207271  0.1122207271
#>  [157,] -1.8587628317 -1.8587628317 -1.8587628317
#>  [158,] -1.5071530601 -1.5071530601 -1.5071530601
#>  [159,]  1.8739990455  1.8739990455  1.8739990455
#>  [160,]  3.4279757551  3.4279757551  3.4279757551
#>  [161,]  0.6638520569  0.6638520569  0.6638520569
#>  [162,]  0.0482050973  0.0482050973  0.0482050973
#>  [163,] -1.8681684237 -1.8681684237 -1.8681684237
#>  [164,]  0.7814157375  0.7814157375  0.7814157375
#>  [165,]  0.5871628678  0.5871628678  0.5871628678
#>  [166,]  1.6588656773  1.6588656773  1.6588656773
#>  [167,]  0.6844394330  0.6844394330  0.6844394330
#>  [168,] -0.1454636125 -0.1454636125 -0.1454636125
#>  [169,]  1.3276212229  1.3276212229  1.3276212229
#>  [170,]  0.0619293020  0.0619293020  0.0619293020
#>  [171,]  1.0790231890  1.0790231890  1.0790231890
#>  [172,] -1.8032760330 -1.8032760330 -1.8032760330
#>  [173,]  1.8690920410  1.8690920410  1.8690920410
#>  [174,]  0.8774310172  0.8774310172  0.8774310172
#>  [175,] -0.9302953803 -0.9302953803 -0.9302953803
#>  [176,] -1.0516304506 -1.0516304506 -1.0516304506
#>  [177,]  0.2979350013  0.2979350013  0.2979350013
#>  [178,]  0.0774877238  0.0774877238  0.0774877238
#>  [179,]  1.3667734939  1.3667734939  1.3667734939
#>  [180,] -2.2378580767 -2.2378580767 -2.2378580767
#>  [181,]  1.1909175856  1.1909175856  1.1909175856
#>  [182,]  0.0974681958  0.0974681958  0.0974681958
#>  [183,]  0.0210380793  0.0210380793  0.0210380793
#>  [184,]  1.2797856428  1.2797856428  1.2797856428
#>  [185,] -0.9273881962 -0.9273881962 -0.9273881962
#>  [186,]  0.1859988516  0.1859988516  0.1859988516
#>  [187,]  2.0667993880  2.0667993880  2.0667993880
#>  [188,]  2.1246689578  2.1246689578  2.1246689578
#>  [189,] -1.2790586279 -1.2790586279 -1.2790586279
#>  [190,] -0.2392797032 -0.2392797032 -0.2392797032
#>  [191,]  1.8106317160  1.8106317160  1.8106317160
#>  [192,] -2.4902887890 -2.4902887890 -2.4902887890
#>  [193,] -2.0283650108 -2.0283650108 -2.0283650108
#>  [194,]  1.2366519827  1.2366519827  1.2366519827
#>  [195,] -0.4649959637 -0.4649959637 -0.4649959637
#>  [196,] -0.8386175992 -0.8386175992 -0.8386175992
#>  [197,]  0.7275400452  0.7275400452  0.7275400452
#>  [198,]  0.5606813444  0.5606813444  0.5606813444
#>  [199,]  1.5484212996  1.5484212996  1.5484212996
#>  [200,]  4.1495994996  4.1495994996  4.1495994996
#>  [201,]  0.3538982774  0.3538982774  0.3538982774
#>  [202,] -3.9271254191 -3.9271254191 -3.9271254191
#>  [203,] -0.8052943021 -0.8052943021 -0.8052943021
#>  [204,]  3.0247217767  3.0247217767  3.0247217767
#>  [205,]  1.3831006767  1.3831006767  1.3831006767
#>  [206,] -2.9330780345 -2.9330780345 -2.9330780345
#>  [207,] -1.1549787549 -1.1549787549 -1.1549787549
#>  [208,] -1.2516783106 -1.2516783106 -1.2516783106
#>  [209,]  1.9514733444  1.9514733444  1.9514733444
#>  [210,]  0.2379130054  0.2379130054  0.2379130054
#>  [211,] -1.1982540108 -1.1982540108 -1.1982540108
#>  [212,]  1.7346093749  1.7346093749  1.7346093749
#>  [213,]  0.6102780971  0.6102780971  0.6102780971
#>  [214,] -0.0169038270 -0.0169038270 -0.0169038270
#>  [215,]  2.6425404146  2.6425404146  2.6425404146
#>  [216,]  0.2319015951  0.2319015951  0.2319015951
#>  [217,]  0.1306587529  0.1306587529  0.1306587529
#>  [218,]  0.2159798566  0.2159798566  0.2159798566
#>  [219,]  1.8741904597  1.8741904597  1.8741904597
#>  [220,] -1.1373311223 -1.1373311223 -1.1373311223
#>  [221,]  0.2867476303  0.2867476303  0.2867476303
#>  [222,] -3.3167651869 -3.3167651869 -3.3167651869
#>  [223,] -0.6472349037 -0.6472349037 -0.6472349037
#>  [224,]  0.8703245665  0.8703245665  0.8703245665
#>  [225,] -0.6981325003 -0.6981325003 -0.6981325003
#>  [226,]  0.5889908768  0.5889908768  0.5889908768
#>  [227,] -2.3865063159 -2.3865063159 -2.3865063159
#>  [228,] -0.2491391371 -0.2491391371 -0.2491391371
#>  [229,]  0.1036564266  0.1036564266  0.1036564266
#>  [230,]  2.0300800181  2.0300800181  2.0300800181
#>  [231,]  2.7195292403  2.7195292403  2.7195292403
#>  [232,] -0.7839505791 -0.7839505791 -0.7839505791
#>  [233,]  1.0316857345  1.0316857345  1.0316857345
#>  [234,]  0.8326247935  0.8326247935  0.8326247935
#>  [235,] -2.7544539369 -2.7544539369 -2.7544539369
#>  [236,] -0.7711489501 -0.7711489501 -0.7711489501
#>  [237,]  2.1222612217  2.1222612217  2.1222612217
#>  [238,]  0.1947234221  0.1947234221  0.1947234221
#>  [239,] -1.1144195291 -1.1144195291 -1.1144195291
#>  [240,]  0.3352564708  0.3352564708  0.3352564708
#>  [241,] -0.6702600546 -0.6702600546 -0.6702600546
#>  [242,]  1.3753831754  1.3753831754  1.3753831754
#>  [243,]  2.2730284290  2.2730284290  2.2730284290
#>  [244,]  2.6645635789  2.6645635789  2.6645635789
#>  [245,] -2.2237440716 -2.2237440716 -2.2237440716
#>  [246,]  0.6643715724  0.6643715724  0.6643715724
#>  [247,]  0.2111992773  0.2111992773  0.2111992773
#>  [248,] -1.2128635438 -1.2128635438 -1.2128635438
#>  [249,] -0.0903205953 -0.0903205953 -0.0903205953
#>  [250,] -1.7705689144 -1.7705689144 -1.7705689144
#>  [251,] -2.7541505384 -2.7541505384 -2.7541505384
#>  [252,] -0.9075087864 -0.9075087864 -0.9075087864
#>  [253,]  0.1386878779  0.1386878779  0.1386878779
#>  [254,]  0.7363067591  0.7363067591  0.7363067591
#>  [255,] -0.4427593319 -0.4427593319 -0.4427593319
#>  [256,] -1.1156702288 -1.1156702288 -1.1156702288
#>  [257,] -0.0153272238 -0.0153272238 -0.0153272238
#>  [258,] -2.3464453668 -2.3464453668 -2.3464453668
#>  [259,]  0.5681734253  0.5681734253  0.5681734253
#>  [260,] -2.4158498868 -2.4158498868 -2.4158498868
#>  [261,]  0.2753182744  0.2753182744  0.2753182744
#>  [262,]  2.4742493806  2.4742493806  2.4742493806
#>  [263,]  0.7486599784  0.7486599784  0.7486599784
#>  [264,]  0.8948575151  0.8948575151  0.8948575151
#>  [265,] -3.5233052578 -3.5233052578 -3.5233052578
#>  [266,] -1.6848420547 -1.6848420547 -1.6848420547
#>  [267,] -0.9318068783 -0.9318068783 -0.9318068783
#>  [268,] -2.8192831706 -2.8192831706 -2.8192831706
#>  [269,]  0.1098819007  0.1098819007  0.1098819007
#>  [270,]  0.6332697203  0.6332697203  0.6332697203
#>  [271,]  0.5732487283  0.5732487283  0.5732487283
#>  [272,]  1.0387559896  1.0387559896  1.0387559896
#>  [273,]  1.4839524940  1.4839524940  1.4839524940
#>  [274,]  0.7543696625  0.7543696625  0.7543696625
#>  [275,]  0.5061546843  0.5061546843  0.5061546843
#>  [276,] -0.9039528692 -0.9039528692 -0.9039528692
#>  [277,] -0.5478363684 -0.5478363684 -0.5478363684
#>  [278,]  0.1166785754  0.1166785754  0.1166785754
#>  [279,] -0.1267028101 -0.1267028101 -0.1267028101
#>  [280,] -1.1185029037 -1.1185029037 -1.1185029037
#>  [281,] -1.0276025089 -1.0276025089 -1.0276025089
#>  [282,] -1.8025445834 -1.8025445834 -1.8025445834
#>  [283,] -3.0241276767 -3.0241276767 -3.0241276767
#>  [284,]  0.8175535069  0.8175535069  0.8175535069
#>  [285,]  0.7345423752  0.7345423752  0.7345423752
#>  [286,]  2.9807136440  2.9807136440  2.9807136440
#>  [287,] -1.7633297314 -1.7633297314 -1.7633297314
#>  [288,]  1.7820967090  1.7820967090  1.7820967090
#>  [289,] -2.6897532533 -2.6897532533 -2.6897532533
#>  [290,] -1.9267193322 -1.9267193322 -1.9267193322
#>  [291,]  1.3765274312  1.3765274312  1.3765274312
#>  [292,] -2.7473072389 -2.7473072389 -2.7473072389
#>  [293,] -2.0981871884 -2.0981871884 -2.0981871884
#>  [294,] -1.3386250578 -1.3386250578 -1.3386250578
#>  [295,] -1.6495026546 -1.6495026546 -1.6495026546
#>  [296,]  0.7271247626  0.7271247626  0.7271247626
#>  [297,] -0.3760531432 -0.3760531432 -0.3760531432
#>  [298,]  3.0079019940  3.0079019940  3.0079019940
#>  [299,]  0.6339680437  0.6339680437  0.6339680437
#>  [300,] -0.1681017287 -0.1681017287 -0.1681017287
#>  [301,]  0.8392188259  0.8392188259  0.8392188259
#>  [302,]  0.0647900542  0.0647900542  0.0647900542
#>  [303,] -0.2865079113 -0.2865079113 -0.2865079113
#>  [304,]  0.3198018829  0.3198018829  0.3198018829
#>  [305,] -1.4872184252 -1.4872184252 -1.4872184252
#>  [306,] -2.8325578658 -2.8325578658 -2.8325578658
#>  [307,]  1.8658286974  1.8658286974  1.8658286974
#>  [308,]  0.2815269294  0.2815269294  0.2815269294
#>  [309,]  2.0614379277  2.0614379277  2.0614379277
#>  [310,] -1.9565021555 -1.9565021555 -1.9565021555
#>  [311,] -1.2696405655 -1.2696405655 -1.2696405655
#>  [312,] -0.0884776849 -0.0884776849 -0.0884776849
#>  [313,] -0.2629123442 -0.2629123442 -0.2629123442
#>  [314,] -0.9475134768 -0.9475134768 -0.9475134768
#>  [315,]  3.5586860512  3.5586860512  3.5586860512
#>  [316,] -0.7679478028 -0.7679478028 -0.7679478028
#>  [317,] -0.7835007519 -0.7835007519 -0.7835007519
#>  [318,] -3.4300716311 -3.4300716311 -3.4300716311
#>  [319,]  0.3701204013  0.3701204013  0.3701204013
#>  [320,]  0.2574144901  0.2574144901  0.2574144901
#>  [321,]  1.9775510498  1.9775510498  1.9775510498
#>  [322,]  1.4516772647  1.4516772647  1.4516772647
#>  [323,]  0.7143802477  0.7143802477  0.7143802477
#>  [324,] -2.5333232208 -2.5333232208 -2.5333232208
#>  [325,]  1.1493114757  1.1493114757  1.1493114757
#>  [326,]  1.1495537560  1.1495537560  1.1495537560
#>  [327,] -0.6777746473 -0.6777746473 -0.6777746473
#>  [328,]  2.0269313180  2.0269313180  2.0269313180
#>  [329,]  2.5526643573  2.5526643573  2.5526643573
#>  [330,]  0.2309324951  0.2309324951  0.2309324951
#>  [331,]  2.5695464927  2.5695464927  2.5695464927
#>  [332,]  1.5448881424  1.5448881424  1.5448881424
#>  [333,] -0.7863377242 -0.7863377242 -0.7863377242
#>  [ reached getOption("max.print") -- omitted 691 rows ]